initial.rev7
reverted back to method of initial.rev5 but not removing class attr any
more as that would mess up when widget was moved to PanelUI-popup and
when in palette (as in not placed). just had to touch up the style
sheet. now supports making it not hoverable/activeable in areas test of:
TabsToolbar, nav-bar, and OTHER which was tested in (PersonalToolbar
[bookmarks] and toolbar-menubar [filed edit view tools etc bar])